# optimize the fixed32 and fixed64 encoding