@stepancheg 2025-09-21
Sorry, this crate is barely supported.

Ownership of the crate name is transferred to authors of protobuf, who are very close to make official release of protobuf4, which uses none of this code.

https://docs.rs/protobuf/4.32.1-release/protobuf/index.html

(This has `-release` suffix, counterintuitively meaning it is not released yet.)